requirements.txt

flask>=2.3.0
python-dotenv>=1.0.0
gunicorn>=20.0.0  # ← Add this line
google-generativeai==0.3.2
